Add files via upload
This was an attempt to explain to my college professors my findings on the proper direction for future teachings in programming Audio. There are tons of mistakes and it is not organized. After watching an interview of a Professor give a presentation about what he would like to teach us in this class they want to develop. So I went home, stewed on it. And this is what I sent my professors. After that moment, I had new found respect from my college professors.